+++ src/services/index.tsx
@@ -1,0 +1,5 @@
+import { getToken, refreshToken } from './authService';
+
+const authService = { getToken, refreshToken };
+
+export { authService };
